LRCI-36 Move stopwatch related logic into BaseBuild and expand stopwatch data reporting ability to all builds.
